Update Block_explorer.md
I was just learning github when i committed theis file. i believe i fixed teh images issues, as well as changed the LINKS in the section.

NOTE: the links are now hyperlinked but only to ark.io since i dont know the end link for the sections i linked to.